fix(mapper4): fix uint8_t overflow causing SIGFPE and wrong CHR pointer
Two integer overflow bugs that broke every large MMC3 game:

1. Kirby's Adventure (512 KB CHR-ROM):
   `uint8_t num_1kb = num_chr_rom * 8` → 64 * 8 = 512 wraps to 0
   → `bank % 0` → SIGFPE (floating-point exception / integer div-by-zero).
   Fixed by widening num_1kb and num_8kb to uint16_t in mapper_004.c
   and the resolve_chr_bank() parameter.

2. SMB3 / Mega Man 3 (256 KB PRG-ROM):
   `uint16_t prg_rom_len = prg_rom_size * 0x4000` → 16 * 0x4000 = 0x40000
   wraps to 0, so chr_rom pointer was calculated as raw_data + 16 (the
   header) instead of raw_data + 16 + 256KB — pointing chr_rom at the
   start of PRG-ROM data, producing scrambled / blank graphics.
   Fixed by widening prg_rom_len and chr_rom_len to uint32_t in
   cartridge.h (correct max is ~4 MB PRG / ~2 MB CHR per iNES spec).

All 27 PPU unit tests pass.

+++ arch/6502/cartridge.h
@@ -74,9 +74,9 @@ struct nes_cartridge {
     uint8_t *trainer;
     uint16_t trainer_len;
     uint8_t *prg_rom;
-    uint16_t prg_rom_len;
+    uint32_t prg_rom_len;  // up to 4 MB (iNES max)
     uint8_t *chr_rom;
-    uint16_t chr_rom_len;
+    uint32_t chr_rom_len;  // up to 2 MB (iNES max)
     uint8_t chr_ram_allocated;  // 1 if chr_rom is malloc'd CHR-RAM, 0 if from file
     uint8_t *pc_inst_rom;
     uint8_t *pc_prom;
